% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_sichuan_heatmap}
\alias{plot_sichuan_heatmap}
\title{绘制四川省人才适配度热力图}
\usage{
plot_sichuan_heatmap(data)
}
\arguments{
\item{data}{数据框，必须包含 "city" (市州名) 和 "score" (匹配度)}
}
\value{
Echarts 交互式地图对象
}
\description{
绘制四川省人才适配度热力图
}
