% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_team_radar}
\alias{plot_team_radar}
\title{绘制团队心理特质雷达图}
\usage{
plot_team_radar(scores)
}
\arguments{
\item{scores}{包含5个分数的向量 (开放性, 责任心, 神经质, 外向性, 宜人性)}
}
\value{
Echarts 交互式雷达图
}
\description{
绘制团队心理特质雷达图
}
